Don’t follow directory symlinks found while walking
A symlink to a directory probably points to a virtual environment or a
dependency or some other directory that users don’t expect us to
traverse, so ignore it unless it was passed directly on the command
line.

We still follow symlinks to directories passed directly, as well as
symlinks to files.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>