Correct EOL offset for lines ending with multi-byte char
For each line, ruff iterates over UTF-32 characters and calculates the
candidate of newline offset by adding 1 to byte-offset of the
characters. This does not work correctly if there is a line which ends
with a multi-byte character.

This commit fixes the issue by adding byte-length of the iterating
characters instead of 1.