Add snapshot tests with inline check reporting
Idea here is to make it easier to visually confirm that the checks are
in the right location. When developing lints, I had to "manually"
transfer the location (row, column) from the existing snapshot to the
original source file fixture.

This also serves as a first step toward better reporting with source
code.