Remove CFormatBytes from vendored code.
Giving up on linting `b"..." % ...` expressions for now, since pyflakes
doesn't do it. However, I think it would be pretty easy to add this as a
ruff-only check in future. The only tricky bit is getting human-readable
messages out of the pottentially horrible bytestrings - maybe we don't
return the invalid specifiers to the user, just point out there is one?