Fix detection of changed imports in isort plugin
With the correct handling of line endings recently introduced, the isort
plugin got into an endless loop because the check whether something was
changed made incorrect assumptions.

Specifically, `dedent` will rejoin the lines using LF. When compared to
the source which may have CRLF or just CR, this condition will never be
true, even if nothing visually actually changed.

An alternative solution would be to reimplement dedent and indent to be
aware of the line endings. Something we eventually might have to do, as
there are still some places that (incorrectly) assume line endings are
always just LF.

Closes #1490