Add isort.force-sort-within-sections setting
This commit is a (hacky) first attempt at addressing issue #1003.

The default `isort` behavior is `force-sort-within-sections = false`,
which places `from X import Y` statements after `import X` statements.

When `force-sort-within-sections = true` all imports are sorted by module name.
When module names are equivalent, the `import` statement comes before the `from` statement.
- https://pycqa.github.io/isort/docs/configuration/options.html#force-sort-within-sections

Initially I tried out an `enum` + `match` approach but I was struggling to come
up with anything that didn't require `Copy` support.

---

The approach in this PR works but is... lacking in clarity.
The `merge_imports` function returns a vector of tuples that represent the final import order:

```
[(0 or 1, idx)]
```

Where...
- `0` means use the `import_block.import` vector with index `idx`
- `1` means use the `import_block.import_from` vector with index `idx`

Those integers are essentially a stand-in for a type or enum,
I'm just not sure what the idiomatic thing to return would be.

- An enum like `ImportType::ImportFrom / ImportType::Import`?
- Is there a better way to do this lazily (eg. return an iterator) rather than eagerly construct the vector in `merge_imports`?