Actually fix wasm-pack build command
I initially attempted to run `wasm-pack build -p ruff` which gave the
error message:

    Error: crate directory is missing a `Cargo.toml` file; is `-p` the wrong directory?

I interpreted that as wasm-pack looking for the "ruff" directory because
I specified -p ruff, however actually the wasm-pack build usage is:

    wasm-pack build [FLAGS] [OPTIONS] <path> <cargo-build-options>

And I was missing the <path> argument. So this actually wasn't at all a
bug in wasm-pack but just a confusing error message. And the symlink
hack I introduced in the previous commit didn't actually work ... I only
accidentally omitted the `-p` when testing (which ended up as `ruff`
being the <path> argument) ... CLIs are fun.