fix: issue D401 only for non-test/property functions and methods
Extend test fixture to verify the targeting.

Includes two "attribute docstrings" which per PEP 257 are not recognized
by the Python bytecode compiler or available as runtime object
attributes. They are not available for us either at time of writing, but
include them for completeness anyway in case they one day are.