Add a smoke test to `ruff_fmt` to debug formatter changes (#4)
Adds a smoke test to `ruff_fmt` to debug formatter changes.

This PR changes the return value of `fmt` to `Result<Formatted>` to allow printing the formatted IR in the quick test. The motivation behind returning `Formatted` is that the IR represents the formatted source code. Printing it only changes its representation from the IR to a string.