Rename new `ruff rule` output format to "pretty"
The new `ruff rule` output format introduced in
551b810aebb86f396e4c45a8b6b4406a8cf8cf71 doesn't print Markdown but
rather some rich text with escape sequences for colors and links,
it's actually the "text" format that prints Markdown, so naming the new
format "markdown" is very confusing. This commit therefore renames it to
"pretty".

This isn't a breaking change since there hasn't been a release yet.