refactor: Spawn process to generate CLI documentation
This PR spawns the `ruff_cli` in the `dev generate-cli-help` command instead of linking against it. The benefit is that it removes the need that `ruff_cli` is a binary and library crate.

My main motivation was that clippy ended up generating a ton of unused warnings when I moved code from `ruff` to `ruff_cli` that only was used in the binary crate. This could be worked around by splitting the code into more modules but that feels like unnecessary extra work.

The main downside of this approach is that it potentially is slower because `cargo` sometimes happens to build the CLI again.