refactor: Replace `Vec` in options metadata with static array
This refactor replaces the usages of a `Vec` inside `get_available_options` with a static array slice. I used this refactor as an exercise to familiarize myself with options.

The old implementation used the `ConfigurationOptions` trait that had a `get_available_options` method to retrieve the child-options of a group. The downside was that `get_available_options` allocated and returned a new vector with each invocation.

This implementation replaces the `ConfigurationOptions` trait with a `OptionsGroup` struct that reference a static option entries slice and implements the `get` operation to find a child option.

## Test Plan

```bash
cargo dev generate-options
```

Generates no changes

```bash
> cargo run --bin ruff -q config flake8-annotations
mypy-init-return
suppress-dummy-args
suppress-none-returning
allow-star-arg-any
ignore-fully-untyped
```