Address review comment to remove unwraps
As suggested I have refactored find_shell_keyword to return a struct
that includes the has_shell and keyword fields. After applying this
refactor I realised that the name of as_shell function no longer really
makes sense as it was just determining whether the value for a given
keyword was truthy so I've refactored that as well.