Fix diagnostic range of (doc-)line-too-long
I noticed in the byte offset refactor that the (doc-)line-too-long
start range can point past the end of the line if the line contains characters with
a width greater or equal to two. This is because we set the start locations' column
to `limit` instead of computing the character at which the line width is exceeded.