Add .git-rewrite folder to default ignored folder paths
The .git-rewrite folder is generated by the git-filter-branch command,
and very unlikely to be something ruff wants to touch.

git-filter-branch is a git command which allows running commands over an
entire git history. ruff is a good candidate for the type of commands
that may be run, so it's sensible to add this in the defaults.

See git-filter-branch(1) for more info