Lint pyproject.toml
This adds a new rule `InvalidPyprojectToml` that lints pyproject.toml by checking if https://github.com/PyO3/pyproject-toml-rs can parse it. This means the linting is currently very basic, e.g. we don't check whether the name is actually a valid python project name or appropriately normalized. It does catch errors e.g. with invalid dependency requirements or problems withs the license specifications. It is open to be extended in the future (validate name, SPDX expressions, classifiers, ...), either in ruff or in pyproject-toml-rs.

TODOs:

 - [ ] Why does `FilePattern::Builtin("pyproject.toml")` alone not work? I've added `*.toml` for now to test but this should be changed before merging (help wanted)
 - [ ] Run this over the ecosystem CI dataset