Create `Short` SerializationFormat `
Rename the `Text` `SerializationFormat` to `Short`, but keep the `Text` alias for now.

The motivation for renaming is that I want to introduce a new `Human` format that pretty prints the output.
*Short* works better with Human because both formats are text based. The amin difference is that one shows
extended information, including source frames, whereas the other only shows a brief summary.

I'm open to other names: An alternatives that I considered is *brief* .